chore: moved DTOs to seperate folder
