+++ server/server/src/main/java/se/experis/tidsbanken/server/DTOs/CommentDTO.java
@@ -1,4 +1,4 @@
-package se.experis.tidsbanken.server.models;
+package se.experis.tidsbanken.server.DTOs;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.springframework.lang.NonNull;
